Create README.md
web-design #practice-01 #description
This is my first attempt at building a personal portfolio website, and I am aware that it is not currently responsive. For the best viewing experience, please make sure you view this project on a laptop. I would greatly appreciate your feedback on my progress so far. Thank you for taking the time to view my website!